Breaking change:
Move rarely used demo-specific dashboard widgets to extensions
(data_layers toggle, covariates model)
