+++ examples/interval_annotations.html
@@ -11,8 +11,8 @@
     <script src="../dist/locuszoom.vendor.min.js" type="text/javascript"></script>
     <script src="../dist/locuszoom.app.min.js" type="text/javascript"></script>
 
-      <!-- Helpers that use the external library to add functionality to LocusZoom -->
-  <script src="../dist/ext/lz-intervals-track.min.js" type="application/javascript"></script>
+    <!-- Helpers that use the external library to add functionality to LocusZoom -->
+    <script src="../dist/ext/lz-intervals-track.min.js" type="application/javascript"></script>
 
     <title>LocusZoom.js ~ Interval Annotations Example</title>
